fix(select-courses): error when store branch fails (#402)
Do not throw the error when storage fails. This caused the load
categories to fail sometimes because of an issue with the store branch
method. Instead of fixing the method for storing the branch, I opted to
just catch and ignore the error, since it occurs so rarely. We do not
really need to store the branch, it is mostly for the convenience and
not needed. Therefore, it will suffice to ignore it for the few users
that are affected. Also, bl-web is legacy, and I did not want to dig up
Angular initialization trees. 🙃