fix: upgrade bl-model to fix login (#410)
The import location for the login fail error changed, but there were
multiple versions of bl-model and it used the wrong one, leading to the
login fail message not showing.

To upgrade bl-model without introducing countless type errors, the dead
code for matching must be deleted. Matching is now handled by bl-next.